Explicitly instantiate templates for View
